cleanup style move Fonts
